% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpl_df.R
\name{dexp_gpl_df}
\alias{dexp_gpl_df}
\title{Create a list of derivatives of expected gpl losses from a \code{gpl_df}}
\usage{
dexp_gpl_df(df, F)
}
\arguments{
\item{df}{a \code{gpl_df}, as created by \code{\link[=new_gpl_df]{new_gpl_df()}}.}

\item{F}{list of predictive cdfs, one per row of \code{df}.}
}
\value{
A list of functions of an allocation \code{x}, one per target.
}
\description{
Create a list of derivatives of expected gpl losses from a \code{gpl_df}
}
\examples{
gdf <- new_gpl_df(N = 2, alpha = 0.5)
dexp_gpl_df(gdf, F = list(pnorm, pnorm))

}
