% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpl_loss.R
\name{dexp_under_loss}
\alias{dexp_under_loss}
\title{Derivative of the expected under-prediction loss}
\usage{
dexp_under_loss(g = "x", dg = NULL, F)
}
\arguments{
\item{g}{a non-decreasing increment function, supplied either as a function
or as a string in the variable \code{x} such as \code{"log(x)"}.}

\item{dg}{derivative of \code{g}; if \code{NULL} or \code{NA} it is derived from \code{g}.}

\item{F}{predictive cdf of the outcome.}
}
\value{
A function of an allocation \code{x}.
}
\description{
Derivative of the expected under-prediction loss
}
\examples{
dexp_under_loss(F = pnorm)(0)

}
