% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpl_loss.R
\name{exp_over_loss}
\alias{exp_over_loss}
\title{Expected g-linear loss for over-prediction of a random outcome}
\usage{
exp_over_loss(dg = function(y) 1, F)
}
\arguments{
\item{dg}{derivative of the increment function \code{g}.}

\item{F}{predictive cdf of the outcome.}
}
\value{
A function of an allocation \code{x} giving the expected g-linear loss of
over-predicting an outcome distributed according to \code{F}.
}
\description{
Expected g-linear loss for over-prediction of a random outcome
}
\examples{
eol <- exp_over_loss(F = pnorm)
eol(0)

}
