% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpl_loss.R
\name{exp_under_loss}
\alias{exp_under_loss}
\title{Expected g-linear loss for under-prediction of a random outcome}
\usage{
exp_under_loss(dg = function(y) 1, F)
}
\arguments{
\item{dg}{derivative of the increment function \code{g}.}

\item{F}{predictive cdf of the outcome.}
}
\value{
A function of an allocation \code{x} giving the expected g-linear loss of
under-predicting an outcome distributed according to \code{F}.
}
\description{
Expected g-linear loss for under-prediction of a random outcome
}
\examples{
eul <- exp_under_loss(F = pnorm)
eul(0)

}
