% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpl_df.R
\name{gpl}
\alias{gpl}
\title{Get the gpl loss data frame of an allocation}
\usage{
gpl(adf)
}
\arguments{
\item{adf}{an object of class \code{allocated}, as returned by \code{\link[=allocate]{allocate()}}.}
}
\value{
The \code{gpl_df} stored in the \code{"gpl_df"} attribute of \code{adf}.
}
\description{
Get the gpl loss data frame of an allocation
}
\examples{
fc <- add_pdqr_funs(
  tibble::tibble(target_names = c("a", "b"), dist = "norm", mean = c(5, 8), sd = 1),
  types = c("p", "q")
)
gpl(allocate(fc, K = 10))

}
