% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpl_loss.R
\name{gpl_loss_fun}
\alias{gpl_loss_fun}
\title{Create a generalized piecewise linear (gpl) loss function}
\usage{
gpl_loss_fun(g = "x", kappa = 1, alpha = NA, O = NA, U = NA, offset = 0)
}
\arguments{
\item{g}{a non-decreasing increment function, supplied either as a function
or as a string in the variable \code{x} such as \code{"log(x)"}.}

\item{kappa}{scale factor.}

\item{alpha}{normalized loss when the outcome \code{y} exceeds the allocation \code{x}.
Exactly one of \code{alpha} and \code{U} must be supplied.}

\item{O}{cost incurred when the allocation \code{x} exceeds the outcome \code{y};
equals \code{kappa * (1 - alpha)}.}

\item{U}{cost incurred when the outcome \code{y} exceeds the allocation \code{x};
equals \code{kappa * alpha}.}

\item{offset}{a constant, or a function of \code{y}, added to the loss. The
default of \code{0} gives a loss with \code{L(x, x) = 0}.}
}
\value{
A function of an allocation \code{x} and an outcome \code{y} giving the loss.
}
\description{
The gpl loss of an allocation \code{x} against an outcome \code{y} is
\deqn{L(x, y) = \kappa[(1 - \alpha)(g(x) - g(y))_+ + \alpha (g(y) - g(x))_+] + \mathrm{offset}(y).}
With \code{g(x) = x} this is the pinball (quantile) loss at level \code{alpha}, scaled
by \code{kappa}. Despite the name it need not be piecewise linear, since \code{g} may
be any non-decreasing increment function.
}
\examples{
# pinball loss at the median
L <- gpl_loss_fun(alpha = 0.5)
L(x = 1, y = 3)

# the equivalent over/under-cost parameterization
L2 <- gpl_loss_fun(O = 0.5, U = 0.5)
L2(x = 1, y = 3)

}
