% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpl_df.R
\name{meb_gpl_df}
\alias{meb_gpl_df}
\title{Create marginal expected benefit functions for a \code{gpl_df}}
\usage{
meb_gpl_df(df, F, w)
}
\arguments{
\item{df}{a \code{gpl_df}, as created by \code{\link[=new_gpl_df]{new_gpl_df()}}.}

\item{F}{list of predictive cdfs, one per row of \code{df}.}

\item{w}{weights (costs per unit allocated) used in the budget constraint.}
}
\value{
A list of functions of an allocation \code{x}, one per target.
}
\description{
The marginal expected benefit of allocating to target \code{i} is
\deqn{\Lambda_i(x) = -\frac{1}{w_i}\frac{d}{dx} E L_i(x, Y_i)
      = \frac{\kappa_i}{w_i} g'(x) (\alpha_i - F_i(x)),}
a decreasing function of \code{x}. \code{\link[=allocate]{allocate()}} equalizes it across targets.
}
\examples{
gdf <- new_gpl_df(N = 2, alpha = 0.5)
meb <- meb_gpl_df(gdf, F = list(pnorm, pnorm), w = 1)
meb[[1]](0)

}
