% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oracle.R
\name{oracle_alloscore_direct}
\alias{oracle_alloscore_direct}
\title{Find and score the oracle allocation directly}
\usage{
oracle_alloscore_direct(y, K, w, gpl_fns)
}
\arguments{
\item{y}{observed outcomes, one per target.}

\item{K}{vector of budgets. Cannot be supplied via \code{df}.}

\item{w}{numeric vector of costs per unit of resource allocated to each
target.}

\item{gpl_fns}{list of gpl loss functions, one per target.}
}
\value{
A tibble with columns \code{oracle} and \code{components_oracle}.
}
\description{
Find and score the oracle allocation directly
}
\examples{
oracle_alloscore_direct(
  y = c(3, 5), K = 4, w = 1,
  gpl_fns = list(gpl_loss_fun(alpha = 1), gpl_loss_fun(alpha = 1))
)

}
