% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpl_loss.R
\name{over_loss}
\alias{over_loss}
\title{Basic g-linear loss for over-prediction of an outcome}
\usage{
over_loss(g = function(u) u)
}
\arguments{
\item{g}{a non-decreasing increment function, supplied either as a function
or as a string in the variable \code{x} such as \code{"log(x)"}.}
}
\value{
A function of an allocation \code{x} and an outcome \code{y} that penalizes
only over-prediction.
}
\description{
Basic g-linear loss for over-prediction of an outcome
}
\examples{
ol <- over_loss()
ol(x = 3, y = 1)
ol(x = 1, y = 3)

}
