% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdize_params.R
\name{stdize_met_params}
\alias{stdize_met_params}
\title{Convert meteorologist parameters to \code{kappa} and \code{alpha}}
\usage{
stdize_met_params(C, L)
}
\arguments{
\item{C}{marginal cost per unit of recommended protection.}

\item{L}{marginal loss due to under-provision of needed resources.}
}
\value{
A data frame with columns \code{kappa} and \code{alpha}.
}
\description{
Convert meteorologist parameters to \code{kappa} and \code{alpha}
}
\examples{
stdize_met_params(C = 2, L = 10)

}
