% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdize_params.R
\name{stdize_ou_params}
\alias{stdize_ou_params}
\title{Convert over/under-prediction costs to \code{kappa} and \code{alpha}}
\usage{
stdize_ou_params(O, U)
}
\arguments{
\item{O}{cost incurred when the allocation exceeds the outcome.}

\item{U}{cost incurred when the outcome exceeds the allocation.}
}
\value{
A data frame with columns \code{kappa} and \code{alpha}.
}
\description{
Convert over/under-prediction costs to \code{kappa} and \code{alpha}
}
\examples{
stdize_ou_params(O = 1, U = 3)

}
