% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpl_loss.R
\name{under_loss}
\alias{under_loss}
\title{Basic g-linear loss for under-prediction of an outcome}
\usage{
under_loss(g = function(u) u)
}
\arguments{
\item{g}{a non-decreasing increment function, supplied either as a function
or as a string in the variable \code{x} such as \code{"log(x)"}.}
}
\value{
A function of an allocation \code{x} and an outcome \code{y} that penalizes
only under-prediction.
}
\description{
Basic g-linear loss for under-prediction of an outcome
}
\examples{
ul <- under_loss()
ul(x = 1, y = 3)
ul(x = 3, y = 1)

}
