% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim.R
\name{weights.allocated}
\alias{weights.allocated}
\title{Get the weights of an allocated data frame}
\usage{
\method{weights}{allocated}(object, ...)
}
\arguments{
\item{object}{an \code{allocated} object, as returned by \code{\link[=allocate]{allocate()}}.}

\item{...}{ignored.}
}
\value{
The named vector of weights used in the allocation.
}
\description{
Get the weights of an allocated data frame
}
\examples{
fc <- add_pdqr_funs(
  tibble::tibble(target_names = c("a", "b"), dist = "norm", mean = c(5, 8), sd = 1),
  types = c("p", "q")
)
weights(allocate(fc, K = 10))

}
