% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloscore2-package.R
\docType{package}
\name{alloscore2-package}
\alias{alloscore2}
\alias{alloscore2-package}
\title{alloscore2: Allocation Scoring Rules for Hubverse Model Output}
\description{
Implements allocation scoring rules for evaluating probabilistic forecasts. Given forecasts for several targets and a shared resource budget, an allocation is computed by minimizing expected generalized piecewise linear loss subject to the budget constraint, and forecasts are scored by the realized loss of that allocation relative to the loss of an allocation made by an oracle that knew the observed outcomes. Model output and observed data are read in the formats used by the hubverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reichlab.github.io/alloscore2/}
  \item \url{https://github.com/reichlab/alloscore2}
  \item Report bugs at \url{https://github.com/reichlab/alloscore2/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Reich \email{nick@umass.edu} (\href{https://orcid.org/0000-0003-3503-9899}{ORCID})

Authors:
\itemize{
  \item Nicholas Reich \email{nick@umass.edu} (\href{https://orcid.org/0000-0003-3503-9899}{ORCID})
  \item Aaron Gerding \email{agerding@umass.edu} (Author of the original alloscore package) [copyright holder]
}

Other contributors:
\itemize{
  \item Consortium of Infectious Disease Modeling Hubs [copyright holder]
}

}
\keyword{internal}
